Refactor quirks registry as a class.
